refactor: BaseResponse 타입을 @shared/types/api에서 @confeti/core/http로 이동 및 관련 코드 수정
