+++ apps/client/src/shared/apis/home/home-queries.ts
@@ -1,8 +1,9 @@
 import { queryOptions } from '@tanstack/react-query';
 
+import { BaseResponse } from '@confeti/core/http';
+
 import { END_POINT } from '@shared/constants/api';
 import { HOME_QUERY_KEY } from '@shared/constants/query-key';
-import { BaseResponse } from '@shared/types/api';
 import {
   CarouselPerformancesResponse,
   SuggestMusicPerformanceResponse,
