+++ apps/client/src/shared/apis/my-history/setlist-queries.ts
@@ -1,9 +1,10 @@
 import { queryOptions } from '@tanstack/react-query';
 
+import { BaseResponse } from '@confeti/core/http';
+
 import { get } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
 import { SETLIST_QUERY_KEY } from '@shared/constants/query-key';
-import { BaseResponse } from '@shared/types/api';
 import {
   ArtistMusicSearchRequest,
   ArtistMusicSearchResponse,
