+++ apps/client/src/shared/apis/search/search-queries.ts
@@ -1,9 +1,10 @@
 import { queryOptions } from '@tanstack/react-query';
 
+import { BaseResponse } from '@confeti/core/http';
+
 import { get } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
 import { SEARCH_QUERY_KEY } from '@shared/constants/query-key';
-import { BaseResponse } from '@shared/types/api';
 import {
   RelatedArtistResponse,
   RelatedPerformanceResponse,
