+++ apps/client/src/shared/apis/timetable/festival-timetable-mutation.ts
@@ -1,6 +1,7 @@
+import { BaseResponse } from '@confeti/core/http';
+
 import { del, patch, post } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
-import { BaseResponse } from '@shared/types/api';
 import {
   FestivalIds,
   FestivalTimetableResponse,
