+++ apps/client/src/shared/apis/timetable/festival-timetable-queries.ts
@@ -1,9 +1,10 @@
 import { infiniteQueryOptions, queryOptions } from '@tanstack/react-query';
 
+import { BaseResponse } from '@confeti/core/http';
+
 import { get } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
 import { FESTIVAL_TIMETABLE_QUERY_KEY } from '@shared/constants/query-key';
-import { BaseResponse } from '@shared/types/api';
 import {
   FestivalTimetableResponse,
   FestivalTimetableResponseExtended,
