--- apps/client/src/shared/types/api.ts
@@ -1,12 +1,0 @@
-export type BaseResponse<T> = {
-  status: number;
-  message: string;
-  data: T;
-};
-
-export interface ErrorResponse {
-  message?: string;
-  code?: number;
-}
-
-export type BaseResponseWithoutData = Omit<BaseResponse<unknown>, 'data'>;
