+++ packages/core/src/http/types.ts
@@ -1,3 +1,16 @@
 import type { AxiosError, InternalAxiosRequestConfig } from 'axios';
 
 export type { AxiosError, InternalAxiosRequestConfig };
+
+export type BaseResponse<T> = {
+  status: number;
+  message: string;
+  data: T;
+};
+
+export interface ErrorResponse {
+  message?: string;
+  code?: number;
+}
+
+export type BaseResponseWithoutData = Omit<BaseResponse<unknown>, 'data'>;
